Update: server.py

+++ fastapi_backend/server.py
@@ -98,7 +98,7 @@ def train_model(learning_rate, num_epochs, dataloader, device, model, opti):
         for i, (images, labels) in enumerate(dataloader):
             # Images and labels to device
             images = images.to(device)
-            labels = labels.to(device)
+            labels = labels.to(device).long()
 
             # Forward pass
             outputs = model(images)
